Minimales Design (#77)
* feat: Improve log display on index page

The code changes include updating the CSS styles for the log container on the index page to improve its display. The font family, margin, and text alignment have been modified. Additionally, a border, padding, and background color have been added to enhance the visual appearance.

* feat: Update subtitle link style on index page

The code changes modify the CSS styles for the subtitle link on the index page. The class "subtitle-link" has been added to the link element to apply specific styles. This improves the visual appearance and enhances the user experience.

* feat: Remove unused FocusOnNavigate component

The code changes remove the unused FocusOnNavigate component from the App.razor file. This component was previously responsible for focusing on the h1 element when navigating, but it is no longer needed. This cleanup improves code readability and reduces unnecessary complexity.

* feat: Update landing page heading to "Willkommen bei Kurmann Videoschnitt"

* feat: Add line-height to hover effect in site.css