Implementierung der manuellen Steuerung des TimerTriggerService (#83)
* feat: Add Kurmann.Videoschnitt.Application.Services namespace to _Imports.razor

* Timer can be manually started and stopped

* feat: Add navigation links for metadata processing and timer service

The code changes in the `Index.razor` file add navigation links for executing the metadata processing and timer service functionalities. Clicking on these links will trigger the corresponding API calls and update the log display accordingly. Additionally, the CSS styles in the `site.css` file have been modified to improve the appearance of the navigation links.

* feat: Add TimerTriggerService as a singleton service

The code changes in the `Program.cs` file add `TimerTriggerService` as a singleton service using the `AddSingleton` method. This ensures that there is only one instance of the `TimerTriggerService` throughout the application.

* feat: Register TimerTriggerService as a singleton and hosted service