Metadatenverarbeitung mit Wolverine-Benachrichtigungsdienst (#112)
* distinct files for message types

* feat: Add MetadataProcessedEventLogHandler class

The code changes add a new class `MetadataProcessedEventLogHandler` in the `Application/Handlers` directory. This class handles the `MetadataProcessedEvent` by sending a log message to the `LogHub` and printing it to the console.

* feat: Add ProcessMetadataRequestHandler class

The code changes add a new class `ProcessMetadataRequestHandler` in the `Features/MetadataProcessor/Handler` directory. This class handles the `ProcessMetadataRequest` by processing the metadata asynchronously using the `MetadataProcessingService`.

* feat: Add ProcessMetadataRequestHandler class

* feat: Remove SignalR dependency in ProcessMetadataRequestHandler

* feat: Update ProcessMetadataRequestHandler to use IHubContext

The code changes update the `ProcessMetadataRequestHandler` class to use `IHubContext` instead of `LogHub` for sending log messages. This change removes the dependency on SignalR and allows for more flexibility in sending messages to clients.

* processed metadata event

* feat: Update OnInitialized method to use async

The code changes update the `OnInitialized` method in the `Index.razor` file to use the `OnInitializedAsync` method instead. This change allows for asynchronous initialization of the component and improves the performance of the page load.