Programmbibliothek für Nachrichten (#121)
* feat: Remove MessageLogHubService from application startup

The code changes remove the registration of the `MessageLogHubService` from the application startup in the `Program.cs` file. This service was previously responsible for logging message requests to the `LogHub` using SignalR. However, it is no longer needed and has been deleted from the codebase.

* feat: Add Messages project as a project reference in Application.csproj

* feat: Remove Engine project reference from Application.csproj

* Moved messages to dedicated assembly

* feat: Add Messages project to the repository

The code changes add a new project, Messages.csproj, to the repository. This project contains the necessary configuration and settings for the Messages module of the application.

* feat: Update project reference to Messages.csproj in MetadataProcessor.csproj

* feat: Remove unused MessageService dependency

The code changes remove the unused `MessageService` dependency from the `Program.cs` file in the `Application` project. This dependency was no longer being used and has been removed to clean up the codebase.

* integrated new messages assembly to event handlers and services