Dateiauflistung für Metadatenverarbeitung (#136)
* Implemented MediaFileListener service

* feat: Log processed media files in MetadataProcessedEventLogHandler

The code changes in `MetadataProcessedEventLogHandler.cs` add the functionality to log the names of processed media files. If there are no processed files, a log message indicating that no files were processed is sent. Otherwise, each processed file name is logged individually. This improves the logging capabilities of the application and provides more detailed information about the processing of media files.

* feat: Add SupportedMediaFilesForMetadataProcessingFoundEventHandler

This commit adds a new event handler, `SupportedMediaFilesForMetadataProcessingFoundEventHandler`, to handle the event when supported media files for metadata processing are found. The handler sends log messages to the `LogHub` using SignalR to inform about the supported media files and their names.