Album-Verarbeitung (#197)
* feat: Add album metadata to FFmpegMetadata class

* file scoped namespace on MetadataProcessorEngine

* feat: Add InfuseMediaLibrarySettings to MetadataProcessorSettings

The `MetadataProcessorSettings` class has been updated to include a new property `InfuseMediaLibrarySettings`, which represents the settings for the Infuse media library. This change allows for better configuration and integration with the Infuse media library.

* feat: Add TargetDirectoryResolver class for resolving target directory

* integrate TargetDirectoryResolver

* feat: Update service lifetimes in ServiceCollectionExtensions

The service lifetimes in the ServiceCollectionExtensions class have been updated from `AddSingleton` to `AddScoped`. This change ensures that the services are properly scoped and managed within the application.

* Added new test data with album quicktime metadata tag inside

* Added generated xml from metdata

* removed ffmpeg raw txt data