Vereinfachte Berechtigungslösung (#222)
* Save xml with FileTransferService as well

* feat: Add overwriteExistingFile parameter to WriteFileAsync method

The `WriteFileAsync` method in the `FileTransferService` class has been updated to include a new `overwriteExistingFile` parameter. This parameter allows the caller to specify whether an existing file should be overwritten when writing the content to the file.

* Implemented TransferPermissionsAsync and TransferDirectoryPermissionsAsync

* Re-integrate .NET file writing methods because we are going to use permission transfers instead

* Use unix commands only for passing permissions, use .net copy and file move

* feat: Implement ClearSpecificPermissionsAsync method in FileTransferService

This commit adds the `ClearSpecificPermissionsAsync` method to the `FileTransferService` class. This method allows for the deletion of specific permissions in a directory, so that the permissions from the directory structure above are inherited. It iterates through all files and subdirectories in the target directory, applying the `chmod` command to set the permissions.

* Remove specific file permission after copying and removing

* wrap arguments

* Removed file transfer service because it causes issues here to be find out why