chore: Update deploy task command to use deploy.sh script
The code changes modify the "publish" task in the `.vscode/tasks.json` file to use the `deploy.sh` script instead of the `dotnet publish` command. This change allows for more flexibility in the deployment process.

Note: The commit message has been generated based on the provided code changes and recent user commits.