Veröffentlichung als systemweiter (vorübergehend) LaunchDaemon auf MacOS (#279)
* Defined external tools settings

* integrate external tools settings to common assembly

* Integrated ffmpeg full path from setting into FFmpegColorConversionService

* Added path for FFProbe tool as well

* Integrated ffmpeg and ffprobe path settings in FFmpegMetadataService

* Integrate SIPS full path from application settigns into MacOSColorConversionService

* Integrate ffmpeg path into constructor instead of method in FFmpegColorConversionService

* feat: Integrate Kestrel HTTP endpoint URL into appsettings.Development.json

The code changes add a new configuration setting for the Kestrel HTTP endpoint URL in the `appsettings.Development.json` file. This allows the application to specify the URL for the Kestrel server to listen on during development.

Note: The commit message has been generated based on the provided code changes and recent user commits.

* feat: Add Kestrel HTTP endpoint URL configuration to appsettings.json

* chore: Update deploy task command to use deploy.sh script

The code changes modify the "publish" task in the `.vscode/tasks.json` file to use the `deploy.sh` script instead of the `dotnet publish` command. This change allows for more flexibility in the deployment process.

Note: The commit message has been generated based on the provided code changes and recent user commits.

* Pass workspace from .vscode/tasks.json to deploy bash script

* Created deploy script to publish application as LaunchDaemons including unload and load

* chore: Add .vscode/deploy.sh to .gitignore as exception

The code changes add the `.vscode/deploy.sh` file to the `.gitignore` list. This ensures that the `deploy.sh` script is  tracked by Git.

* Do not include debug files on deployed application

* Do not include full namespace on assembly names

* Shortened application name adjusted on launch.json

* chore: Ignore production appsettings.json in .gitignore

The code changes add the "appsettings.Production.json" file to the ".gitignore" list. This ensures that the production app settings file is not tracked by Git.

Note: The commit message has been generated based on the provided code changes and recent user commits.

* Logging level is same on all environments, only detailed errors is on development env

* Removed infuse banner file postfix from appsettings.json

* New paths and sub directories defined in the deploy.sh

* added filename as settings property

* Use proper launch.json and not launchSettings and make sure its actually useful e.g. auto launch brower.

* Default port is already 5000

* Use host default builder. Not necessary to add appsettings files and env vars manually

* deploy Library/LaunchDaemon plist as well and copy appsettings as well

* chore: Update deploy.sh to publish the .NET application as a Single-File without Debug-Files

The deploy.sh script has been modified to publish the .NET application as a Single-File without including Debug-Files. This change improves the deployment process by reducing the size of the published application and removing unnecessary files.

Note: The commit message has been generated based on the provided code changes and recent user commits.

* manual copy of appsettings json files not necessary since we use Host.CreateDefaultBuilder

* Do not put Application into Application directory because of "doppelt gemoppelt"