feat: Ignore unsupported files in InputDirectoryReaderService
This commit adds functionality to the InputDirectoryReaderService to ignore files that are not supported image or video files. When a file is not supported, it is logged as an information message and added to the list of ignored files. This ensures that only supported files are processed further.

Note: The commit message has been generated based on the provided code changes and recent user commits.