HeathCheck prüft auch die SIPS-Version (#329)
* feat: Refactor HealthCheckWorkflow to log FFmpeg and Sips versions

* feat: Refactor RunHealthCheck method to return HealthCheckResponse

This commit refactors the `RunHealthCheck` method in the `Engine` class to return a `HealthCheckResponse` object instead of a `Result<List<string>>`. The `HealthCheckResponse` object contains the versions of FFmpeg and SIPS. This change improves the clarity and structure of the code.

* feat: Update ToolsVersionService to include GetSipsVersion method

This commit adds a new method `GetSipsVersion` to the `ToolsVersionService` class. This method executes the command `sips --version` and returns the version of SIPS, a tool used for image processing on macOS. This enhancement improves the functionality of the service by providing the ability to retrieve the SIPS version.