MetadatenProcessor mit MediaSetOrganizer ersetzt (Umbenennung) (#340)
* Add media set directory path settings

* print media set path settings

* Renamed MetadataProcessor to MediaSetOrganizer

* feat: Refactor logging in MediaSetService and MediaTypeDetectorService

The code changes in `MediaSetService.cs` and `MediaTypeDetectorService.cs` refactor the logging statements. The unnecessary log message in `MediaSetService.cs` has been removed, and the log message in `MediaTypeDetectorService.cs` now uses string interpolation to include the file name. This improves the clarity and consistency of the logging in the code.