Eingangsverzeichnis kann auch ein Verzeichnis sein, dass gerade Dateien exportiert (#345)
* feat: Fix null reference exception in MediaIntegratorService

* feat: Check for supported files before organizing into media sets

The code changes in `MediaSetService.cs` add a check to ensure that there are supported image or video files in the input directory before attempting to organize them into media sets. If no supported files are found, a warning message is logged and an empty list of media sets is returned. This improves the efficiency of the media set organization process and prevents unnecessary processing of unsupported files.

* feat: Check for supported files before organizing into media sets

* feat: Add check for supported files before organizing into media sets