feat: Add support for including subdirectories in input directory reader service
The code changes in `InputDirectoryReaderService.cs` modify the `ReadInputDirectoryAsync` method to include an optional parameter `includeSubdirectories`. This allows the user to specify whether or not to search for files in subdirectories of the input directory. By default, subdirectories are included in the search. This enhancement provides more flexibility and control over the file search process.