Diverse kleine Verbesserung wie bspw. die Erkennung von Tilde als Homeverzeichnis (#352)
* Move final cut pro export files to sub directory as test setting

* feat: Add support for including subdirectories in input directory reader service

The code changes in `InputDirectoryReaderService.cs` modify the `ReadInputDirectoryAsync` method to include an optional parameter `includeSubdirectories`. This allows the user to specify whether or not to search for files in subdirectories of the input directory. By default, subdirectories are included in the search. This enhancement provides more flexibility and control over the file search process.

* deploy and publish are the same task content here, so removed deploy and renamed shell scrip accordingly

* ExpandHomeDirectory (interpreting tilde)

* Removed jpg testfiles

* feat: Entnehme konvertierte JPG-Dateien für Poster und Fanarts

* add portrait mode test image