feat: Organize videos by codec and support recursive search
This commit adds functionality to the `organize_videos.py` script to organize videos by codec and support recursive search. Previously, the script only organized MOV files by codec in the specified directory. Now, it will also skip files that are already in one of the target directories. Additionally, if the `-r` flag is provided, the script will recursively search all subdirectories. This enhancement improves the flexibility and usability of the script.