feat: Add delete_file function to file_utils.py
This commit adds a new function, delete_file, to the file_utils.py script. The delete_file function is responsible for deleting the specified file. It checks if the file exists and then removes it using the os.remove() function. If the file does not exist, it prints a message indicating that the file was not found. This new function enhances the functionality of the file_utils.py script by providing a convenient way to delete files.